+++ ts/packages/core/src/index.ts
@@ -4,6 +4,7 @@ export { OpenAIProvider } from './provider/OpenAIProvider';
 export { ComposioProvider } from './provider/ComposioProvider';
 export { BaseNonAgenticProvider, BaseAgenticProvider } from './provider/BaseProvider';
 export type { BaseComposioProvider } from './provider/BaseProvider';
+export { ToolRouter } from './models/ToolRouter.experimental';
 export { jsonSchemaToZodSchema, removeNonRequiredProperties } from './utils/jsonSchema';
 export { AuthScheme } from './models/AuthScheme';
 export { MCP } from './models/MCP';
@@ -20,6 +21,7 @@ export * from './types/triggers.types';
 export * from './types/mcp.types';
 export * from './types/files.types';
 export * from './types/connectionRequest.types';
+export * from './types/toolRouter.types';
 export * as constants from './utils/constants';
 
 export { default as logger } from './utils/logger';